Ensure slave nodes use same host as CT
https://github.com/erlang/otp/blob/bc22321f10c1ad71fb7e28275a8e6ed26a34d4ab/lib/stdlib/src/slave.erl#L287-L309

The above code checks to see if the slave is running on the same host as the master, and if not, uses `rsh` to configure the slave. Since `rsh` is blocked on my machine, I got connection refused.

So, I decided that we should use whatever the Common Test nodes hostname is, which will probably be `localhost`. By using the same hostname, we should remove the chance that `rsh` will be used.