Avoid double notifying on node termiation.
When a node was terminated aten would sometimes send two notifications
to registered listeners. One when it detected the node as not responding
and one next interval when it detected it as gone.

This adds a check for gone case to see if the node has already been
detected as down and only sends a down notification in this scenario
if the prior state is not down.

Also return `ignore` if trying to monitor the current node.